+++ packages/sdk/src/funcs/customerCreditsCreateAccount.ts
@@ -20,6 +20,7 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
+import * as errors from "../models/errors/index.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import { APICall, APIPromise } from "../types/async.js";
@@ -38,6 +39,7 @@ export function customerCreditsCreateAccount(
 ): APIPromise<
   Result<
     components.AccountResponseDto,
+    | errors.CustomerCreditsErrorResponseDto
     | CreemError
     | ResponseValidationError
     | ConnectionError
@@ -63,6 +65,7 @@ async function $do(
   [
     Result<
       components.AccountResponseDto,
+      | errors.CustomerCreditsErrorResponseDto
       | CreemError
       | ResponseValidationError
       | ConnectionError
@@ -139,8 +142,13 @@ async function $do(
   }
   const response = doResult.value;
 
+  const responseFields = {
+    HttpMeta: { Response: response, Request: req },
+  };
+
   const [result] = await M.match<
     components.AccountResponseDto,
+    | errors.CustomerCreditsErrorResponseDto
     | CreemError
     | ResponseValidationError
     | ConnectionError
@@ -151,9 +159,10 @@ async function $do(
     | SDKValidationError
   >(
     M.json(201, components.AccountResponseDto$inboundSchema),
+    M.jsonErr(409, errors.CustomerCreditsErrorResponseDto$inboundSchema),
     M.fail([400, 401, 404, "4XX"]),
     M.fail("5XX"),
-  )(response, req);
+  )(response, req, { extraFields: responseFields });
   if (!result.ok) {
     return [result, { status: "complete", request: req, response }];
   }